wip: working on moving actor html to actor
Updated action paths to:
actor.<actor>.action.<action>.<account>

These are all defined in bla/actor.go now to keep them in one place.

Next subject is
actor.<actor>.render

This will take a request and return a reply, which contains
templated html to show on the page.
There was a hacky test for this without NATS in htmx.go